consolidate some factory girl helpers and definitions
